reorganize FRLG RNG programs
