+++ SerialPrograms/Source/Pokemon/Pokemon_AdvRng.h
@@ -15,13 +15,13 @@
 namespace PokemonAutomation{
 namespace Pokemon{
 
-struct IvGroup{
+struct AdvIvGroup{
     uint8_t iv0;
     uint8_t iv1;
     uint8_t iv2;
 };
 
-struct IVs{
+struct AdvIVs{
     uint8_t hp = 0;
     uint8_t attack = 0;
     uint8_t defense = 0;
@@ -30,7 +30,7 @@ struct IVs{
     uint8_t speed = 0;
 };
 
-enum class Nature{
+enum class AdvNature{
     Hardy,
     Lonely,
     Brave,
@@ -59,26 +59,26 @@ enum class Nature{
     Any
 };
 
-enum class Gender{
+enum class AdvGender{
     Male,
     Female,
     Any
 };
 
-enum class Ability{
+enum class AdvAbility{
     Zero,
     One,
     Any
 };
 
-enum class ShinyType{
+enum class AdvShinyType{
     Normal,
     Star,
     Square,
     Any
 };
 
-enum class RngMethod{
+enum class AdvRngMethod{
     Method1,
     Method2,
     Method4,
@@ -88,7 +88,7 @@ enum class RngMethod{
 struct AdvRngState{
     uint16_t seed;
     uint64_t advance;
-    RngMethod method;
+    AdvRngMethod method;
     uint32_t s0;
     uint32_t s1;
     uint32_t s2;
@@ -103,18 +103,28 @@ struct AdvRngState{
 struct AdvPokemonResult{
     uint32_t pid;
     uint8_t gender;
-    Nature nature;
-    Ability ability;
-    IVs ivs;
+    AdvNature nature;
+    AdvAbility ability;
+    AdvIVs ivs;
+};
+
+struct AdvObservedPokemon{
+    AdvGender gender;
+    AdvNature nature;
+    AdvAbility ability;
+    std::vector<uint8_t> level;
+    std::vector<StatReads> stats;
+    std::vector<EVs> evs;
+    AdvShinyType shiny;
 };
 
 struct AdvRngFilters{
-    Gender gender;
-    Nature nature;
-    Ability ability;
+    AdvGender gender;
+    AdvNature nature;
+    AdvAbility ability;
     IvRanges ivs;
-    ShinyType shiny;
-    RngMethod method;
+    AdvShinyType shiny;
+    AdvRngMethod method;
 };
 
 class AdvRng{
@@ -123,7 +133,7 @@ public:
     AdvRngState state;
 
     AdvRng(uint16_t seed, AdvRngState state);
-    AdvRng(uint16_t seed, uint64_t min_advances, RngMethod method = RngMethod::Method1);
+    AdvRng(uint16_t seed, uint64_t min_advances, AdvRngMethod method = AdvRngMethod::Method1);
 
     void set_seed(uint16_t seed);
     void set_state_advances(uint64_t advances);
