+++ SerialPrograms/Source/PokemonFRLG/Inference/PokemonFRLG_BattleLevelUpReader.h
@@ -12,6 +12,7 @@
 #include "Common/Cpp/Color.h"
 #include "CommonFramework/ImageTools/ImageBoxes.h"
 #include "CommonFramework/Language.h"
+#include "Pokemon/Pokemon_StatsCalculation.h"
 
 
 namespace PokemonAutomation{
@@ -23,14 +24,8 @@ class VideoOverlaySet;
 namespace NintendoSwitch{
 namespace PokemonFRLG{
 
-struct PokemonFRLG_LevelUpStats{
-    std::optional<unsigned> hp;
-    std::optional<unsigned> attack;
-    std::optional<unsigned> defense;
-    std::optional<unsigned> sp_attack;
-    std::optional<unsigned> sp_defense;
-    std::optional<unsigned> speed;
-};
+using namespace Pokemon;
+
 
 class BattleLevelUpReader {
 public:
@@ -38,7 +33,7 @@ public:
 
     void make_overlays(VideoOverlaySet &items) const;
 
-    PokemonFRLG_LevelUpStats read_stats(Logger &logger, const ImageViewRGB32& frame);
+    StatReads read_stats(Logger &logger, const ImageViewRGB32& frame) const;
 
 private:
     Color m_color;
